Implemented login backend & changed proxy
Login button now fetches data from the server and checks if it is valid. It loads into the home page if valid. The proxy in package.json has been set to http://localhost:4000 so that you can connect to the backend server with fetch() without having to type in local host