Implement login frontend
Sorry I forgot to make a message when I merged token-login-stuff into the main branch. Please look at the previous commit for actual file changes. Here is my commit message from token-login-stuff: Make sign up and login webpages actually log in a person. Add authentication hooks and context, which check if a token exists or not. If the token exists, then the user state in the authentication context will be set to include the token. For the navbar and webpages, certain things can be displayed depending on whether a token exists. For example, I made the navbar show sign up and login only if logged out, and create project, logout and profile only if logged in. To use the authentication context, add import { useAuthContext } from '../hooks/use-auth-context' and do const { user } = useAuthContext(). If user is null, the user is logged out. If the user is not null, the user is logged in. The use-login, use-logout, and use-signup hooks are used to login, logout, and signup respectively. Their usage can be seen in Login.js, Sign-up.js, and Navbar.js. As we are using auth-context.js instead of user-context.jsfor the tokens, I removed the tokens state in user-context.js. user-context.js currently has no purpose but may be useful in the future.