Sign-ups & Remove node_modules/
Read advice online that node_modules is redudant if we already commit
the package.json file. because it is inevitable that some changes will
happen, it can slow down git push/pulls because of how massive it is.
that is why letting npm handle dependencies with package.json is a
better idea.