Protected routes and update profile page
Updated app.js to force users back to the login page if they are signed out and want to access other links. Users cannot access login and sign up if they are logged in. When the url is …/profile, the webpage loads the user’s own profile page. Updated backend and frontend routes to achieve this. Added loading context to authContextProvider for loading screens.