Updated profile to show posts dynamically and changed user model
Added a profile context and related components and hooks to render the profile page. Added options for logos in index.html. Updated post-model to include whether or not the project has been completed. Updated the user model to not include an array of post ids anymore, as we can find the posts of a user through the post database.